--- css/fonts.css
@@ -1,73 +1,0 @@
-@font-face {
-  font-family: 'fontawesome';
-  src:  url('../fonts/fontawesome.eot?5aa35k');
-  src:  url('../fonts/fontawesome.eot?5aa35k#iefix') format('embedded-opentype'),
-    url('../fonts/fontawesome.ttf?5aa35k') format('truetype'),
-    url('../fonts/fontawesome.woff?5aa35k') format('woff'),
-    url('../fonts/fontawesome.svg?5aa35k#icomoon') format('svg');
-  font-weight: normal;
-  font-style: normal;
-}
-
-[class^="icon-"], [class*=" icon-"] {
-  /* use !important to prevent issues with browser extensions that change fonts */
-  font-family: 'fontawesome' !important;
-  speak: none;
-  font-style: normal;
-  font-weight: normal;
-  font-variant: normal;
-  text-transform: none;
-  line-height: 1;
-  font-size: 20px;
-
-  /* Better Font Rendering =========== */
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.icon-eye:before {
-  content: "\e903";
-}
-.icon-cogs:before {
-  content: "\e900";
-}
-span.icon-cogs{
-	position: absolute;
-	top: 25px;
-	z-index: 1;
-	right: 5px;
-	color: #ddd;
-	font-size: 28px;
-	opacity: 0.5;
-	transition: 1s;
-}
-span.icon-cogs:hover {
-	text-shadow: 0 0 2px #fff;
-	opacity: 1;
-}
-.icon-cog:before {
-  content: "\e901";
-}
-.icon-lock:before {
-  content: "\e904";
-}
-.icon-address-book:before {
-  content: "\e907";
-}
-.icon-user-circle-o:before {
-  content: "\e905";
-}
-.icon-plus:before {
-  content: "\e902";
-}
-span.icon-plus{
-	position: absolute;
-	top: 60px;
-	z-index: 1;
-	right: 7px;
-	color: #ddd;
-	font-size: 28px;
-}
-span.icon-plus:hover {
-	text-shadow: 0 0 2px #fff;
-}
