+++ package.json
@@ -1,0 +1,42 @@
+{
+  "name": "piping-ssh-web",
+  "version": "0.1.0-SNAPSHOT",
+  "description": "SSH over HTTPS via Piping Server",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build0": "java -jar closure-compiler.jar --js_output_file=js/combinedLibs.comb.js src/defines.js js/src/*.js js/*.js '!**.comb.js'",
+    "build": "webpack --mode development",
+    "serve": "webpack-cli serve --mode development",
+    "serve0": "watch-build-serve -i=.git -i=.idea -i=./node_modules -i=js/combinedLibs.comb.js --public=. --build-command='npm run build'"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nwtgck/piping-ssh-web.git"
+  },
+  "author": "Ryo Ota <nwtgck@nwtgck.org> (https://github.com/nwtgck)",
+  "contributors": [
+    {
+      "name": "Stuart Brooks"
+    }
+  ],
+  "license": "MIT",
+  "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-preset-es2015": "^6.24.1",
+    "copy-webpack-plugin": "^9.0.0",
+    "html-webpack-plugin": "^5.3.1",
+    "ts-loader": "^9.2.3",
+    "typescript": "^4.3.3",
+    "webpack": "^5.39.1",
+    "webpack-cli": "^4.7.2",
+    "webpack-dev-server": "^3.11.2"
+  },
+  "bugs": {
+    "url": "https://github.com/nwtgck/piping-ssh-web/issues"
+  },
+  "homepage": "https://github.com/nwtgck/piping-ssh-web#readme",
+  "dependencies": {
+    "xterm": "^3.1.0"
+  }
+}
